requirements.txt
cfscrape
PySocks